+++ test/support/cases/case.ex
@@ -1,9 +1,21 @@
 defmodule ZTD.Tests.Support.Case do
   use ExUnit.CaseTemplate
 
+  @moduledoc """
+  Default Test Case with important aliases/imports
+  """
+
 
   using do
     quote do
+      alias ZTD.Repo
+      alias ZTD.Tests.Support
+
+      alias Ecto.Query
+      alias Ecto.Changeset
+
+      require Query
+      import Support.Schema
     end
   end
